Move to golang cmd pattern
